requirements.txt
numpy
pandas==1.3.5
matplotlib
seaborn
pytz
pyyaml
shapely
OSMPythonTools
geojson
geopy
geographiclib
